Add rtw_core
