+++ src/chrono_clock.rs
@@ -1,5 +1,6 @@
+use crate::rtw_core::clock::{Clock, Time};
+use crate::rtw_core::datetimew::DateTimeW;
 use chrono::{Date, Datelike, Duration, Local};
-use rtw::{Clock, DateTimeW, Time};
 
 pub struct ChronoClock {}
